requirements.txt
requests
web3
python-telegram-bot[job-queue]==20.5
